feat : 완성인가?

+++ src/main/java/com/ssafy/buyhouse/domain/estate/dto/response/OwnedHouseListResponseDto.java
@@ -1,6 +1,5 @@
 package com.ssafy.buyhouse.domain.estate.dto.response;
 
-import com.ssafy.buyhouse.domain.estate.domain.OwnedHouse;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,9 +16,9 @@ import java.util.stream.Collectors;
 public class OwnedHouseListResponseDto {
 
     private List<OwnedHouseResponseDto> ownedHouseList;
-    private Integer totalOwnedPrice;  // 총 취득가격
-    private Integer totalCurrentPrice; // 총 평가가격
-    private Integer totalPriceDifference; // 총 취득가 대비 증감맥
+    private Long totalOwnedPrice;  // 총 취득가격
+    private Long totalCurrentPrice; // 총 평가가격
+    private Long totalPriceDifference; // 총 취득가 대비 증감맥
     private Double meanPriceDifferenceRate; // 총 취득가 대비 증감율
 
     public static OwnedHouseListResponseDto from(List<OwnedHouseResponseDto> houseList) {
@@ -40,9 +39,9 @@ public class OwnedHouseListResponseDto {
 
         return OwnedHouseListResponseDto.builder()
                 .ownedHouseList(houseList)
-                .totalOwnedPrice((int) statsOwned.getSum())
-                .totalCurrentPrice((int) statsCurrent.getSum())
-                .totalPriceDifference((int) statsDiff.getSum())
+                .totalOwnedPrice(statsOwned.getSum())
+                .totalCurrentPrice(statsCurrent.getSum())
+                .totalPriceDifference(statsDiff.getSum())
                 .meanPriceDifferenceRate(averageRate)
                 .build();
     }
